refactor: move method location
