+++ src/main/java/co/orange/ddanzi/service/auth/AuthService.java
@@ -3,6 +3,7 @@ package co.orange.ddanzi.service.auth;
 import co.orange.ddanzi.domain.user.Authentication;
 import co.orange.ddanzi.domain.user.User;
 import co.orange.ddanzi.domain.user.enums.UserStatus;
+import co.orange.ddanzi.dto.auth.RefreshTokenResponseDto;
 import co.orange.ddanzi.dto.auth.SigninResponseDto;
 import co.orange.ddanzi.dto.auth.VerifyRequestDto;
 import co.orange.ddanzi.dto.auth.VerifyResponseDto;
@@ -17,6 +18,7 @@ import co.orange.ddanzi.service.ItemService;
 import co.orange.ddanzi.service.OrderService;
 import co.orange.ddanzi.service.PaymentService;
 import co.orange.ddanzi.service.TermService;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -82,6 +84,26 @@ public class AuthService {
                 .status(user.getStatus())
                 .build();
         return ApiResponse.onSuccess(Success.CREATE_AUTHENTICATION_SUCCESS, responseDto);
+    }
+
+    @Transactional
+    public ApiResponse<?> refreshAccessToken(String refreshToken) throws JsonProcessingException {
+        if (refreshToken == null || refreshToken.isEmpty()) {
+            return ApiResponse.onFailure(Error.REFRESH_TOKEN_IS_NULL, Map.of("refreshtoken", refreshToken));
+        }
+
+        String email = jwtUtils.getIdFromRefreshToken(refreshToken);
+
+        if (!jwtUtils.isValidRefreshToken(email, refreshToken)) {
+            return ApiResponse.onFailure(Error.REFRESH_TOKEN_EXPIRED, Map.of("refreshtoken", refreshToken));
+        }
+
+        RefreshTokenResponseDto responseDto = RefreshTokenResponseDto.builder()
+                .accesstoken(jwtUtils.createAccessToken(email))
+                .refreshtoken(jwtUtils.createRefreshToken(email))
+                .build();
+
+        return ApiResponse.onSuccess(Success.REFRESH_ACCESS_TOKEN_SUCCESS, responseDto);
     }
 
     @Transactional
